Move FeeError to crate root error module
